Use Dataset factory in client
It seems likely we are going to be pulling multiple sources from the
same repository. This may be true for both DSpace and Dataverse (and
maybe others). We'll also likely need to be able to handle the mapping
from the source to our data model differently for each dataset. The
client shouldn't be tied to the source, but rather to the underlying
repository.

This shifts specific decisions about how to create a dataset outside of
the client so that a Dataverse client, for example, can be used to
harvest multiple datasets.