Add author lookup service
This adds a relatively basic author lookup service for querying the data
warehouse. We'll need to think about how to parse a name to get
something useful out of it. For now, I have a placeholder that assumes
the name will be in the form "First Middle Last" which is obviously not
going to work.

The service will return an iterable of matches. A match is a tuple
containing the name string that was queried and a dictionary-like object
with kerb, full_name and orcid fields.

The service needs a name repository to query. In this case, there is a
data warehouse repository. The repository will need the underlying
database connection configured. To use:

  from hoard.names.db import engine
  from hoard.names.repository import Warehouse
  from hoard.names.service import AuthorService

  engine.configure("oracle://<username>:<password>@<server>:1521/<sid>")
  svc = AuthorService(Warehouse(engine()))
  for author in svc.find("H Abelson"):
    ...