Configure logging
I'm using structlog (https://www.structlog.org) which gives us a bit
better output for less work. If you need a logger do:

  logger = structlog.get_logger()

at module level and use just like you would a stdlib logger.

This also updates the dependency for smart_open which, for some reason,
changed the extras_require for aws to s3 when it got updated to 2.2.