Add author lookup command to CLI
This adds an `author` command to the CLI which can be used to lookup one
or more authors in the data warehouse. The intent is primarily for this
to be used to check how well our name parsing is working and to see what
kind of data we are getting back from the warehouse.

The command will accept any number of names as arguments, but remember
you'll need to quote them:

  hoard author "Foo, Bar" "Foo, Baz"

It will also read from stdin:

  hoard author < author_list.txt