Add S3 logging for name lookup
This configures logging to record the result of author name lookups in
S3. In fact, it will write to any file, including locally. A new option
has been added to the main ingest command called `--name-log`. To write
to an S3 file use a file path like s3://bucket/key.

The logging functionality is all applied through two decorators. As this
feature, theoretically, is temporary it should be easy to remove it when
needed.

You may note that I am not using structlog for this. structlog does not
support different formatting for different loggers (as far as I can
tell). We'll almost certainly want to have more control over formatting
so I have fallen back to stdlib logging for this. I have also disabled
propagation to keep the two logging strategies separate.

Internally, the logs are written to an in-memory buffer until the
command is done and then written to whatever file was specified.