Add mypy stub for smart_open
This adds a typed stub for smart_open. It was mostly done just to see
how this would work. Since only one function is used from this module it
was easy to add just enough type constraints to turn on type checking
for it.

I also added a couple type constraints on the two logging decorators.